+++ flutter/test/sentry_native_channel_test.dart
@@ -20,8 +20,11 @@ void main() {
 
     test('fetchNativeAppStart', () async {
       final map = <String, dynamic>{
+        'pluginRegistrationTime': 1,
         'appStartTime': 0.1,
         'isColdStart': true,
+        // ignore: inference_failure_on_collection_literal
+        'nativeSpanTimes': {},
       };
       final future = Future.value(map);